refactor(login.js): 이벤트 핸들러 함수 구조 변경
