+++ controllers/login.js
@@ -1,4 +1,4 @@
-import { isEmpty, isValidEmail, isValidPassword } from "../modules/auth.js";
+import { isEmpty, isValidEmail, isValidPassword } from "../modules/validate.js";
 
 const emailInput = document.querySelector("#email");
 const passwordInput = document.querySelector("#password");
@@ -7,40 +7,57 @@ const submitBtn = document.querySelector(".form__submit-btn");
 const loginErrorMsg = document.createElement("span");
 loginErrorMsg.classList.add("error-msg");
 emailInput.after(loginErrorMsg);
+
 const passwordErrorMsg = document.createElement("span");
 passwordErrorMsg.classList.add("error-msg");
-passwordInput.after(passwordErrorMsg);
+passwordInput.parentNode.after(passwordErrorMsg);
 
 emailInput.addEventListener("focusout", (e) => {
   const input = e.currentTarget;
   const { value } = input;
 
-  if (isEmpty(value)) {
+  const handleEmpty = () => {
     input.classList.add("error");
     loginErrorMsg.textContent = "이메일을 입력해주세요";
-  } else if (!isValidEmail(value)) {
+  };
+
+  const handleInvalidEmail = () => {
     input.classList.add("error");
     loginErrorMsg.textContent = "잘못된 이메일입니다";
-  } else {
+  };
+
+  const handleValidEmail = () => {
     input.classList.remove("error");
     loginErrorMsg.textContent = "";
-  }
+  };
+
+  if (isEmpty(value)) handleEmpty();
+  else if (!isValidEmail(value)) handleInvalidEmail();
+  else handleValidEmail();
 });
 
 passwordInput.addEventListener("focusout", (e) => {
   const input = e.currentTarget;
   const { value } = input;
 
-  if (isEmpty(value)) {
+  const handleEmpty = () => {
     input.classList.add("error");
     passwordErrorMsg.textContent = "비밀번호를 입력해주세요";
-  } else if (!isValidPassword(value)) {
+  };
+
+  const handleInvalidPassword = () => {
     input.classList.add("error");
     passwordErrorMsg.textContent = "비밀번호를 8자 이상 입력해주세요";
-  } else {
+  };
+
+  const handleValidPassword = () => {
     input.classList.remove("error");
     passwordErrorMsg.textContent = "";
-  }
+  };
+
+  if (isEmpty(value)) handleEmpty();
+  else if (!isValidPassword(value)) handleInvalidPassword();
+  else handleValidPassword();
 });
 
 function updateSignUpButton() {
